refactor(generation-options): Changed to array of Alphabets to simplify logic
