+++ src/app/styles.ts
@@ -1,0 +1,11 @@
+import styled from '@emotion/styled'
+
+export const InputGroup = styled.div`
+  display: block;
+`
+
+export const Container = styled.div`
+  max-width: 550px;
+  margin: 0 auto;
+  padding: 0 20px 20px 20px;
+`
